# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.CreateDialectRequest:
    properties:
      audio_url:
        type: string
      city:
        type: string
      content:
        type: string
      description:
        type: string
      dialect_type:
        type: string
      duration:
        type: integer
      file_size:
        type: integer
      format:
        type: string
      province:
        type: string
      region:
        type: string
      tags:
        type: string
      title:
        type: string
    required:
    - audio_url
    - region
    - title
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UpdateDialectRequest:
    properties:
      city:
        type: string
      content:
        type: string
      description:
        type: string
      dialect_type:
        type: string
      province:
        type: string
      region:
        type: string
      tags:
        type: string
      title:
        type: string
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cntunyuan.org
    name: CNtunyuan Team
    url: https://github.com/Snowitty-Re/CNtunyuan
  description: 团圆寻亲志愿者系统后端 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Snowitty-Re/CNtunyuan
  title: 团圆寻亲志愿者系统 API
  version: 1.0.0
paths:
...
  /dialects:
    get:
      consumes:
      - application/json
      description: 分页获取方言列表，支持关键词、地区、状态等筛选，需要登录权限
      parameters:
      - default: 1
        description: 页码，默认为1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 地区筛选
        in: query
        name: region
        type: string
      - description: 省份筛选
        in: query
        name: province
        type: string
      - description: 城市筛选
        in: query
        name: city
        type: string
      - description: 方言类型
        in: query
        name: type
        type: string
      - description: 状态筛选
        in: query
        name: status
        type: string
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方式(asc/desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectListResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取方言列表
      tags:
      - 方言管理
    post:
      consumes:
      - application/json
      description: 创建新的方言语音记录，需要登录权限
      parameters:
      - description: 创建方言请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.CreateDialectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建方言
      tags:
      - 方言管理
  /dialects/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的方言，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除方言
      tags:
      - 方言管理
    get:
      consumes:
      - application/json
      description: 根据ID获取方言详细信息，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取方言详情
      tags:
      - 方言管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的方言信息，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新方言请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UpdateDialectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新方言
      tags:
      - 方言管理
  /dialects/{id}/comments:
    get:
      consumes:
      - application/json
      description: 获取指定方言的评论列表，支持分页，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: 页码，默认为1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectCommentListResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取方言评论列表
      tags:
      - 方言管理
    post:
      consumes:
      - application/json
      description: 为指定方言添加评论，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      - description: 评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.CreateDialectCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 评论成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectCommentResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 添加方言评论
      tags:
      - 方言管理
  /dialects/{id}/feature:
    delete:
      consumes:
      - application/json
      description: 取消指定方言的精选标记，需要管理员权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 取消方言精选
      tags:
      - 方言管理
    post:
      consumes:
      - application/json
      description: 将指定方言标记为精选内容，需要管理员权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 将方言设为精选
      tags:
      - 方言管理
  /dialects/{id}/like:
    delete:
      consumes:
      - application/json
      description: 取消对指定方言的点赞，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取消点赞成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误或未点赞
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 取消点赞方言
      tags:
      - 方言管理
    post:
      consumes:
      - application/json
      description: 为指定方言点赞，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 点赞成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误或已点赞
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 方言不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 点赞方言
      tags:
      - 方言管理
  /dialects/{id}/play:
    post:
      consumes:
      - application/json
      description: 记录用户播放方言的次数，需要登录权限
      parameters:
      - description: 方言ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 记录成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 记录方言播放
      tags:
      - 方言管理
...
  /dialects/featured:
    get:
      consumes:
      - application/json
      description: 分页获取标记为精选的方言列表，需要登录权限
      parameters:
      - default: 1
        description: 页码，默认为1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.DialectListResponse'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取精选方言列表
      tags:
      - 方言管理
...
securityDefinitions:
  Bearer:
    description: 请输入 JWT Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"